- add stat family of libsys functions
	- apply stat to init, ls and rm

	issue: 145

+++ init/cmd/rm.c
@@ -1,4 +1,5 @@
 #include <sys/errno.h>
+#include <sys/stat.h>
 #include <string.h>
 #include <stdio.h>
 #include <unistd.h>
@@ -7,9 +8,24 @@
 
 /* local functions */
 static int exec(int argc, char **argv){
+	stat_t f_stat;
+
+
 	if(argc < 2){
 		printf("usage: %s <file>\n", argv[0]);
 		return -1;
+	}
+
+	if(stat(argv[1], &f_stat) != 0){
+		printf("file not found\n");
+		return -1;
+	}
+
+	if(f_stat.type == FT_DIR){
+		printf("%s is a directory, do you want to delete it?\n", argv[1]);
+
+		if(fgetc(stdin) != 'y')
+			return 0;
 	}
 
 	if(unlink(argv[1]) != 0){
